requirements.txt
fastapi==0.87.0
matplotlib==3.7.2
numpy==1.23.4
tensorflow-cpu
uvicorn==0.20.0
deta


